LayoutTests/fast/js/sputnik/Conformance/15_Native_ECMA_Script_Objects/15.1_The_Global_Object/S15.1_A1_T2-expected.txt

S15.1_A1_T2

PASS TypeError: Result of expression 'this' [[object DOMWindow]] is not a constructor.

TEST COMPLETE
